feat: refactor + auth
